REFACTOR: moved caller code into agent_utils
