+++ lib/screens/common_widgets/agentic_ui_features/ai_ui_designer/generate_ui_dialog.dart
@@ -1,10 +1,10 @@
 import 'package:apidash/consts.dart';
 import 'package:apidash/providers/collection_providers.dart';
+import 'package:apidash/utils/agent_utils.dart';
 import 'package:apidash/widgets/widget_sending.dart';
 import 'package:apidash_design_system/apidash_design_system.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'ai_ui_designer_agentcalls.dart';
 import 'framework_selector.dart';
 import 'sdui_preview.dart';
