+++ lib/utils/agent_utils.dart
